backend/pyproject.toml
[project]
name = "open-gov-access-backend"
version = "0.1.0"
description = "Backend API for SF Board of Supervisors votes and meeting data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pdfplumber>=0.10.0",
    "python-dateutil>=2.8.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "pymupdf4llm>=0.1.9",
    "selenium>=4.38.0",
    "pymupdf-layout>=1.26.6",
    "google-generativeai>=0.8.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "etl", "models", "scrapers"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

